# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Sport Api Swagger Documentation Anatation",
        "description": "Sport Api Documentation to build the frontend of this site",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "email": "yurayur1990@gmail.com"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "https://sport.napaautomotive.uz/",
            "description": "Sport api server link"
        },
        {
            "url": "https://sport.napaautomotive.uz/",
            "description": "Sport api server link"
        }
    ],
    "paths": {
...
        "/api/allScore": {
            "get": {
                "tags": [
                    "AllScore"
                ],
                "summary": "Get all data from fillials database",
                "description": "Via this link All fillials` datas come",
                "operationId": "allScore",
                "parameters": [
                    {
                        "name": "Accept-Language",
                        "in": "header",
                        "description": "Set language parameter by typing uz, ru, en",
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "400": {
                        "description": "Invalid status value"
                    }
                }
            },
            "post": {
                "tags": [
                    "AllScore"
                ],
                "summary": "Register a allScore",
                "operationId": "storeAllScore",
                "parameters": [
                    {
                        "name": "Accept-Language",
                        "in": "header",
                        "description": "Set language parameter by typing uz, ru, en",
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "requestBody": {
                    "description": "Input data format",
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "direction_id": {
                                        "description": "direction ID",
                                        "type": "integer"
                                    },
                                    "user_id": {
                                        "description": "Direction category ID",
                                        "type": "integer"
                                    },
                                    "all_score": {
                                        "description": "Direction category score",
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "successful operation"
                    }
                }
            }
        },
        "/api/allScore/{allScore}": {
            "get": {
                "tags": [
                    "AllScore"
                ],
                "summary": "Get one data from showAllScore database",
                "description": "Via this link a showAllScore`s data comes to show",
                "operationId": "showAllScore",
                "parameters": [
                    {
                        "name": "Accept-Language",
                        "in": "header",
                        "description": "Set language parameter by typing uz, ru, en",
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "allScore",
                        "in": "path",
                        "description": "ID for avatar data",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "400": {
                        "description": "Invalid status value"
                    }
                }
            }
        },
        "/api/allScore/{allScore}/update": {
            "put": {
                "tags": [
                    "AllScore"
                ],
                "summary": "Updated",
                "description": "Update this AllScore",
                "operationId": "updateAllScore",
                "parameters": [
                    {
                        "name": "Accept-Language",
                        "in": "header",
                        "description": "Set language parameter by typing uz, ru, en",
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "allScore",
                        "in": "path",
                        "description": "allScore that to be updated",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "requestBody": {
                    "description": "Input data format",
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "all_score": {
                                        "description": "Direction category score",
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "400": {
                        "description": "Invalid avatar supplied"
                    },
                    "404": {
                        "description": "avatar not found"
                    },
                    "200": {
                        "description": "successful operation"
                    }
                }
            }
        },
        "/api/allScore/{allScore}/delete": {
            "delete": {
                "tags": [
                    "AllScore"
                ],
                "summary": "Deletes a allScore",
                "operationId": "destroyAllScore",
                "parameters": [
                    {
                        "name": "Accept-Language",
                        "in": "header",
                        "description": "Set language parameter by typing uz, ru, en",
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "allScore",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation"
                    },
                    "400": {
                        "description": "Invalid course supplied"
                    },
                    "404": {
                        "description": "Avatar not found"
                    }
                }
            }
        },
        "/api/allScoreWithUserId/{user_id}": {
            "get": {
                "tags": [
                    "AllScore"
                ],
                "summary": "Get one data from showAllScore database",
                "description": "Via this link a showAllScore`s data comes to show",
                "operationId": "userScore",
                "parameters": [
                    {
                        "name": "Accept-Language",
                        "in": "header",
                        "description": "Set language parameter by typing uz, ru, en",
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "user_id",
                        "in": "path",
                        "description": "ID for avatar data",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "400": {
                        "description": "Invalid status value"
                    }
                }
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearer_token": {
                "type": "apiKey",
                "description": "Enter token in format (Bearer <token>)",
                "name": "Authorization",
                "in": "header"
            }
        }
    }
}